backend/requirements.txt
autopep8==2.3.2
channels==4.2.2
Django==5.2.1
djangorestframework==3.16.0
daphne==4.2.0
